page loading controlled within correct file; Food/Meal class class just call correct methods from DefaultLoader class
