+++ lib/foods.js
@@ -1,40 +1,21 @@
 // This is the base url, last character is a "/" so don't include as first character
 // of specific herokuUrl api request enpoint
 // make requests with `herokuUrl + "api/v1/foods"` etc
-const herokuUrl = require("./heroku-url")
-const CruddyFood = require("./cruddy-food")
-const HtmlEvents = require("./html-events")
-const Filter = require("./filter")
+const DefaultLoader = require("./default-loader")
 const $ = require("jQuery")
 
-class DefaultFoodLoader {
+class Food {
   constructor() {
-    this.crud  = new CruddyFood()
-    this.html  = new HtmlEvents()
-    this.filter = new Filter()
-  }
-
-  getFoods(thisPage) {
-    $('#newFoodForm').hide()
-    $.ajax({
-      type: "GET",
-      url: `${herokuUrl()}api/v1/foods`,
-      success: function(posts) {
-        thisPage.launchPage(posts)
-      }
-    })
-  }
-
-  launchPage(posts) {
-    this.html.fillTable(posts)
-    this.filter.startListener("foods", this.filter)
-    this.crud.startListener()
+    this.loadPage = function() {
+      $(document).ready(function() {
+        const loadPage = new DefaultLoader()
+        loadPage.getFoods(loadPage)
+        })
+    }
   }
 }
 
-$(document).ready(function() {
-  const loadPage = new DefaultFoodLoader()
-  loadPage.getFoods(loadPage)
-})
+const food = new Food()
+food.loadPage()
 
-module.exports = DefaultFoodLoader
+module.exports = Food
